Rename component/movies to component/discover
